+++ src/modules/tools/login/utils/utilsAxios.js
@@ -11,7 +11,9 @@ function addInterceptor (token, interceptorUrlRegex) {
     // Request interceptors for API calls
     axios.interceptors.request.use(
         config => {
-            if (!config.url?.startsWith("http") || (interceptorUrlRegex && config.url?.match(interceptorUrlRegex))) {
+            const configUrl = typeof config.url === "object" ? config.url.origin : config.url;
+
+            if (!configUrl?.startsWith("http") || (interceptorUrlRegex && configUrl?.match(interceptorUrlRegex))) {
                 config.headers.Authorization = `Bearer ${token}`;
             }
             return config;