+++ CHANGELOG.md
@@ -26,6 +26,7 @@
     - In 3D the visible layer list contains the 3D layer now.
 - Login:
     - Fixed token revocation not working correctly
+- OAF-Layer: if parameter 'bbox' ist not defined at layer, 'bbox' in parameter 'datasets' or maps extent is used. If parameter 'bboxCrs' is not set, 'crs' is used.
 
 ---