+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 [Semantic versioning](https://semver.org/spec/v2.0.0.html) is used.
 
-## Unreleased - in development
+## v2.49.0 - 04.09.2024
 ### __Breaking Changes__
 
 ### Added